+++ auth/package.json
@@ -4,7 +4,15 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "start": "ts-node-dev src/index.ts"
+    "start": "ts-node-dev src/index.ts",
+    "test": "jest --watchAll --no-cache"
+  },
+  "jest": {
+    "preset": "ts-jest",
+    "testEnvironment": "node",
+    "setupFilesAfterEnv": [
+      "./src/test/setup.ts"
+    ]
   },
   "keywords": [],
   "author": "",
@@ -15,6 +23,7 @@
     "@types/express": "^4.17.12",
     "@types/jsonwebtoken": "^8.5.5",
     "@types/mongoose": "^5.11.97",
+    "@types/bcryptjs": "^2.4.2",
     "bcrypt": "^5.0.1",
     "bcryptjs": "^2.4.3",
     "cookie-session": "^1.4.0",
@@ -28,7 +37,12 @@
     "typescript": "^4.2.4"
   },
   "devDependencies": {
-    "@types/bcryptjs": "^2.4.2",
-    "prettier": "^2.3.2"
+    "@types/jest": "^27.0.1",
+    "@types/supertest": "^2.0.11",
+    "jest": "^27.1.1",
+    "mongodb-memory-server": "^7.4.0",
+    "prettier": "^2.3.2",
+    "supertest": "^6.1.6",
+    "ts-jest": "^27.0.5"
   }
 }
