+++ auth/src/app.ts
@@ -1,0 +1,32 @@
+import express from "express";
+import "express-async-errors";
+import { json } from "body-parser";
+import cookieSession from "cookie-session";
+
+import { currentUserRouter } from "./routes/currentUser";
+import { signinRouter } from "./routes/signin";
+import { signoutRouter } from "./routes/signout";
+import { signupRouter } from "./routes/signup";
+import { errorHandler } from "./middlewares/errorHandler";
+import { NotFoundError } from "./errors/notFoundError";
+
+const app = express();
+app.set("trust proxy", true);
+app.use(json());
+app.use(
+  cookieSession({
+    name: "session",
+    signed: false,
+    secure: true,
+  })
+);
+
+app.use(currentUserRouter, signinRouter, signoutRouter, signupRouter);
+
+app.all("*", async (req, res) => {
+  throw new NotFoundError();
+});
+
+app.use(errorHandler);
+
+export { app };
