+++ auth/src/index.ts
@@ -1,34 +1,5 @@
-import express from "express";
-import "express-async-errors";
-import { json } from "body-parser";
 import mongoose from "mongoose";
-import cookieSession from "cookie-session";
-
-import { currentUserRouter } from "./routes/currentUser";
-import { signinRouter } from "./routes/signin";
-import { signoutRouter } from "./routes/signout";
-import { signupRouter } from "./routes/signup";
-import { errorHandler } from "./middlewares/errorHandler";
-import { NotFoundError } from "./errors/notFoundError";
-
-const app = express();
-app.set("trust proxy", true);
-app.use(json());
-app.use(
-  cookieSession({
-    name: "session",
-    signed: false,
-    secure: true,
-  })
-);
-
-app.use(currentUserRouter, signinRouter, signoutRouter, signupRouter);
-
-app.all("*", async (req, res) => {
-  throw new NotFoundError();
-});
-
-app.use(errorHandler);
+import { app } from "./app";
 
 const start = async () => {
   if (!process.env.JWT_KEY) {
